+++ template/lugia/class/view/components/MainTable.js
@@ -78,7 +78,7 @@ class MainTable extends PureComponent {
 						maxHeight: '400px',
 					},
 					dropdownMatchSelectWidth: false,
-					getPopupContainer: () => document.getElementById('$PageNameTable'),
+					getPopupContainer: () => document.getElementById('$PageNameMainTable'),
 				},
 			},
 			{
@@ -100,7 +100,7 @@ class MainTable extends PureComponent {
 							});
 						}
 					},
-					getCalendarContainer: () => document.getElementById('$PageNameTable'),
+					getCalendarContainer: () => document.getElementById('$PageNameMainTable'),
 				},
 			},
 			{
@@ -121,7 +121,7 @@ class MainTable extends PureComponent {
 						maxHeight: '400px',
 					},
 					dropdownMatchSelectWidth: false,
-					getPopupContainer: () => document.getElementById('$PageNameTable'),
+					getPopupContainer: () => document.getElementById('$PageNameMainTable'),
 				},
 			},
 		];
@@ -405,7 +405,7 @@ class MainTable extends PureComponent {
 		return (
 			<Fragment>
 				{/* 查询表单 */}
-				<div id='$PageNameTable' style={{ padding: '0px 20px', position: 'relative' }}>
+				<div id='$PageNameMainTable' style={{ padding: '0px 20px', position: 'relative' }}>
 					<SearchForm
 						formItem={formItems}
 						labelSize={'70px'}