+++ template/lugia/hooks/view/components/MainTable.js
@@ -72,7 +72,7 @@ const MainTable = (props) => {
 					maxHeight: '400px',
 				},
 				dropdownMatchSelectWidth: false,
-				getPopupContainer: () => document.getElementById('$PageNameTable'),
+				getPopupContainer: () => document.getElementById('$PageNameMainTable'),
 			},
 		},
 		{
@@ -94,7 +94,7 @@ const MainTable = (props) => {
 						});
 					}
 				},
-				getCalendarContainer: () => document.getElementById('$PageNameTable'),
+				getCalendarContainer: () => document.getElementById('$PageNameMainTable'),
 			},
 		},
 		{
@@ -115,7 +115,7 @@ const MainTable = (props) => {
 					maxHeight: '400px',
 				},
 				dropdownMatchSelectWidth: false,
-				getPopupContainer: () => document.getElementById('$PageNameTable'),
+				getPopupContainer: () => document.getElementById('$PageNameMainTable'),
 			},
 		},
 	];
@@ -393,7 +393,7 @@ const MainTable = (props) => {
 	return (
 		<Fragment>
 			{/* 查询表单 */}
-			<div id='$PageNameTable' style={{ padding: '0px 20px', position: 'relative' }}>
+			<div id='$PageNameMainTable' style={{ padding: '0px 20px', position: 'relative' }}>
 				<SearchForm
 					refs={(ref) => {
 						searchForm.current = ref;