+++ template/redux/class/view/components/MainTable.js
@@ -68,7 +68,7 @@ class MainTable extends PureComponent {
 						maxHeight: '400px',
 					},
 					dropdownMatchSelectWidth: false,
-					getPopupContainer: () => document.getElementById('$PageNameTable'),
+					getPopupContainer: () => document.getElementById('$PageNameMainTable'),
 				},
 			},
 			{
@@ -90,7 +90,7 @@ class MainTable extends PureComponent {
 							});
 						}
 					},
-					getCalendarContainer: () => document.getElementById('$PageNameTable'),
+					getCalendarContainer: () => document.getElementById('$PageNameMainTable'),
 				},
 			},
 			{
@@ -111,7 +111,7 @@ class MainTable extends PureComponent {
 						maxHeight: '400px',
 					},
 					dropdownMatchSelectWidth: false,
-					getPopupContainer: () => document.getElementById('$PageNameTable'),
+					getPopupContainer: () => document.getElementById('$PageNameMainTable'),
 				},
 			},
 		];
@@ -379,7 +379,7 @@ class MainTable extends PureComponent {
 		return (
 			<Fragment>
 				{/* 查询表单 */}
-				<div id='$PageNameTable' style={{ padding: '0px 20px', position: 'relative' }}>
+				<div id='$PageNameMainTable' style={{ padding: '0px 20px', position: 'relative' }}>
 					<SearchForm
 						formItem={formItems}
 						labelSize={'70px'}